+++ src/lib/IExecDealModule.d.ts
@@ -1,6 +1,8 @@
-import IExecConfig from './IExecConfig';
-import { Observable } from '../common/utils/reactive';
-import IExecModule from './IExecModule';
+export * from '../common/types.js';
+
+import IExecConfig from './IExecConfig.js';
+import { Observable } from '../common/utils/reactive.js';
+import IExecModule from './IExecModule.js';
 import {
   Address,
   Addressish,
@@ -13,7 +15,7 @@ import {
   Taskid,
   TaskIndex,
   TxHash,
-} from './types';
+} from '../common/types.js';
 
 declare class DealObservable extends Observable {
   /**